Changed stock to inventario and added confirmation for deletion with views

+++ sevirun/products/views.py
@@ -230,7 +230,7 @@ def delete_product(request, product_id):
         product.save()
         messages.success(request, f'El producto "{product_name}" ha sido eliminado correctamente.')
         return redirect('products')
-    return render(request, 'products/product_confirm_delete.html', {'product': product})
+    return render(request, 'products/product_confirm_delete.html', {'object': product, 'object_name': 'producto', 'cancel_url': 'products'})
 
 @staff_member_required(login_url='login')
 def catalog_management(request):
@@ -312,55 +312,102 @@ def create_colour(request):
 def delete_brand(request, brand_id):
     brand = get_object_or_404(Brand, pk=brand_id)
     if(len(brand.models.all()) > 0):
-        messages.error(request, 'Existen modelos de producto que son de esta marca!')
+        messages.error(request, 'Existen modelos de productos que son de esta marca!')
         return redirect(f"{reverse('catalog_management')}?tab=brands")
-    brand.delete()
-    return redirect(f"{reverse('catalog_management')}?tab=brands")
+    
+    if request.method == 'POST':
+        brand_name = brand.name
+        brand.delete()
+        messages.success(request, f'La marca "{brand_name}" ha sido eliminado correctamente.')
+        return redirect(f"{reverse('catalog_management')}?tab=brands")
+    
+    return render(request, 'products/product_confirm_delete.html', {'object': brand, 'object_name': 'marca', 'cancel_target': f"{reverse('catalog_management')}?tab=brands"})
+
+@staff_member_required(login_url='login')
+def delete_product(request, product_id):
+    product = get_object_or_404(Product, pk=product_id)
+    if request.method == 'POST':
+        product_name = product.name
+        product.is_deleted = True
+        product.save()
+        messages.success(request, f'El producto "{product_name}" ha sido eliminado correctamente.')
+        return redirect('products')
+    return render(request, 'products/product_confirm_delete.html', {'object': product, 'object_name': 'producto', 'cancel_url': 'products'})
 
 @staff_member_required(login_url='login')
 def delete_model(request, model_id):
     model = get_object_or_404(ProductModel, pk=model_id)
-    if(len(model.products) > 0):
+    if Product.objects.filter(model=model).exists():
         messages.error(request, 'Existen productos que son de este modelo!')
         return redirect(f"{reverse('catalog_management')}?tab=models")
-    model.delete()
-    return redirect(f"{reverse('catalog_management')}?tab=models")
+    
+    if request.method == 'POST':
+        model_name = model.name
+        model.delete()
+        messages.success(request, f'El modelo "{model_name}" ha sido eliminado correctamente.')
+        return redirect(f"{reverse('catalog_management')}?tab=models")
+    
+    return render(request, 'products/product_confirm_delete.html', {'object': model, 'object_name': 'modelo', 'cancel_target': f"{reverse('catalog_management')}?tab=models"})
 
 @staff_member_required(login_url='login')
 def delete_type(request, type_id):
     type = get_object_or_404(ProductType, pk=type_id)
-    if(len(type.products) > 0):
+    if Product.objects.filter(type=type).exists():
         messages.error(request, 'Existen productos que son de este tipo!')
         return redirect(f"{reverse('catalog_management')}?tab=types")
-    type.delete()
-    return redirect(f"{reverse('catalog_management')}?tab=types")
+    
+    if request.method == 'POST':
+        type_name = type.name
+        type.delete()
+        messages.success(request, f'El tipo "{type_name}" ha sido eliminado correctamente.')
+        return redirect(f"{reverse('catalog_management')}?tab=types")
+    
+    return render(request, 'products/product_confirm_delete.html', {'object': type, 'object_name': 'tipo', 'cancel_target': f"{reverse('catalog_management')}?tab=types"})
 
 @staff_member_required(login_url='login')
 def delete_material(request, material_id):
     material = get_object_or_404(ProductMaterial, pk=material_id)
-    if(len(material.products) > 0):
+    if Product.objects.filter(material=material).exists():
         messages.error(request, 'Existen productos que usan este material!')
         return redirect(f"{reverse('catalog_management')}?tab=materials")
-    material.delete()
-    return redirect(f"{reverse('catalog_management')}?tab=materials")
+    
+    if request.method == 'POST':
+        material_name = material.name
+        material.delete()
+        messages.success(request, f'El material "{material_name}" ha sido eliminado correctamente.')
+        return redirect(f"{reverse('catalog_management')}?tab=materials")
+    
+    return render(request, 'products/product_confirm_delete.html', {'object': material, 'object_name': 'material', 'cancel_target': f"{reverse('catalog_management')}?tab=materials"})
 
 @staff_member_required(login_url='login')
 def delete_size(request, size_id):
     size = get_object_or_404(ProductSize, pk=size_id)
-    if(size.product_count > 0):
+    if Product.objects.filter(productstock__size=size).exists():
         messages.error(request, 'Existen productos que usan esta talla!')
         return redirect(f"{reverse('catalog_management')}?tab=sizes")
-    size.delete()
-    return redirect(f"{reverse('catalog_management')}?tab=sizes")
+    
+    if request.method == 'POST':
+        size_name = size.name
+        size.delete()
+        messages.success(request, f'La talla "{size_name}" ha sido eliminado correctamente.')
+        return redirect(f"{reverse('catalog_management')}?tab=sizes")
+    
+    return render(request, 'products/product_confirm_delete.html', {'object': size, 'object_name': 'talla', 'cancel_target': f"{reverse('catalog_management')}?tab=sizes"})
 
 @staff_member_required(login_url='login')
 def delete_colour(request, colour_id):    
     colour = get_object_or_404(ProductColour, pk=colour_id)
-    if(colour.product_count > 0):
+    if Product.objects.filter(productstock__colour=colour).exists():
         messages.error(request, 'Existen productos que usan este color!')
         return redirect(f"{reverse('catalog_management')}?tab=colours")
-    colour.delete()
-    return redirect(f"{reverse('catalog_management')}?tab=colours")
+    
+    if request.method == 'POST':
+        colour_name = colour.name
+        colour.delete()
+        messages.success(request, f'El color "{colour_name}" ha sido eliminado correctamente.')
+        return redirect(f"{reverse('catalog_management')}?tab=colours")
+    
+    return render(request, 'products/product_confirm_delete.html', {'object': colour, 'object_name': 'color', 'cancel_target': f"{reverse('catalog_management')}?tab=colours"})
     
 @staff_member_required(login_url='login')
 def edit_brand(request, brand_id):
@@ -487,7 +534,7 @@ def product_stock_view(request):
     prev_page = request.GET.get('from', '/')
 
     return render(request, 'products/product_stock_view.html', {
-        'stocks': ProductStock.objects.all(),
+        'stocks': ProductStock.objects.all().filter(product__is_deleted=False),
         'from': prev_page
     })
 
@@ -544,5 +591,8 @@ def edit_stock(request, stock_id):
 @staff_member_required(login_url='login')
 def delete_stock(request, stock_id):
     stock = get_object_or_404(ProductStock, pk=stock_id)
-    stock.delete()
-    return redirect('product_stock_view')
+    if request.method == 'POST':
+        stock.delete()
+        messages.success(request, f'Inventario del producto eliminado correctamente.')
+        return redirect('product_stock_view')
+    return render(request, 'products/product_confirm_delete.html', {'object': stock, 'object_name': 'inventario del producto', 'cancel_url': 'product_stock_view'})
